refactor(test classes): Изменено расположение 2 тестовых классов
Тесты TestBuilderBankAccount и TestBuyer перенесены в соответствующие директории.
Отредактирован файл README.md.
