+++ README.md
@@ -2,14 +2,37 @@
 
 ## Описание проекта
 
-Этот проект является демонстрацией различных Java-паттернов проектирования. 
+Этот проект является демонстрацией различных Java-паттернов проектирования.
 Проект использует Maven для управления зависимостями и сборкой, а также JUnit для тестирования.
+
 Каждый паттерн находится в своей отдельной директории по пути src/main/java/org/ChiniakinD/patterns.
-Тесты для каждого паттерна находятся в одноименных директориях по пути  src/test/java/org/ChiniakinD.
+Тесты для каждого паттерна находятся в одноименных директориях по пути src/test/java/org/ChiniakinD.
+
+## Зависимости
+
+Для сборки и запуска проекта необходимы следующие зависимости:
+
+- **JDK 17**: Проект использует возможности, доступные в JDK 17.
+- **Maven**: Maven используется для управления зависимостями и сборкой проекта.
 
 ## В проекте представлены следующие паттерны
-### builder 
-### singleton 
-### strategy 
-### prototype 
-### facade 
+
+### Builder: 
+Паттерн для пошагового создания сложных объектов.
+### Singleton:
+Паттерн для обеспечения единственного экземпляра класса.
+### Strategy:
+Паттерн, который определяет семейство схожих алгоритмов и помещает каждый из них в собственный класс, 
+после чего алгоритмы можно взаимозаменять прямо во время исполнения программы.
+### Facade:
+Паттерн для упрощения работы со сложными системами, предоставляет интерфейс взаимодействия.
+### Prototype:
+Паттерн для создания новых объектов путем копирования уже существующих.
+### Chain Of responsibility:
+Паттерн, который позволяет передавать запросы последовательно по цепочке обработчиков.
+
+## Тестирование
+Для запуска тестирования в корневой директории проекта выполните команду:
+**mvn test**
+
+
